MAJOR code reorganization
All init happens in `application.py` for both the web app and the Celery
worker.

Modify the way we run `gunicorn` and `celery` processes.

Move web app from `__init__.py` to `web.py`.

Methods `year_simhash` and `timestamp_simhash` used by the web interface
to return simhash results are moved from the `Discover` Celery class to
independent methods which are called by the UI.

Parameter validation for `year_simhash` and `timestamp_simhash` should
happen in the web app (as early as possible) and not inside the functions.

Unit tests no longer require a YAML conf to run. `run_tests.sh` is not
needed and its removed.

Rename `tests_first.py` to `tests_web.py` and change the way we call the
web app to run the unit tests.