Optimise Celery worker CPU and memory usage
Write all timestamp -> Simhash data to Redis in a batch using Redis
Pipeline. This way, we avoid multiple Redis commnications while processing and
we speed up the whole simhash calculation.

Rename snapshots to captures because captures is the "official" web
archiving term.

Stop requesting JSON and request plain text instead. This makes
downloading and processing CDX faster.

Update job status in the main loop and not inside the threads, this way
we avoid inconsistencies (e.g. one thread write "20 out of 100" and the
other writes "10 out of 100" after that because it happened to be slower.

Use BeautifulSoup decompose method to reduce memory usage.
https://stackoverflow.com/questions/11284643/python-high-memory-usage-with-beautifulsoup