Fetch only the initial part of the data to speed up operation
Get the initial 1MB of the data. There is no point in getting too much
and most of web page HTML is less than 1MB. This limit is effective when
trying to calculate changes for binary files which doesn't work anyway.
The worker doesn't download the whole file which may be huge and it
fails early when the content isn't `text/html`.